requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
celery[redis]>=5.4.0
playwright>=1.49.0
pixelmatch>=0.3.0
Pillow>=11.0.0
numpy>=2.0.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.0
boto3>=1.35.0
